requirements.txt
python-telegram-bot
selenium
redis
logging